Changed the data types of constant fields.

+++ src/main/java/com/api/management/leave/leavemanagementapi/dto/HourConversionDto.java
@@ -5,11 +5,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class HourConversionDto {
     private int hour;
-    private double equivalentDay;
+    private BigDecimal equivalentDay;
 }
